+++ src/sync_queue.c
@@ -3,14 +3,14 @@
 #include<stdio.h>
 #include<stdlib.h>
 
-sync_queue* new_sync_queue(unsigned long long int size, int is_bounded)
+sync_queue* new_sync_queue(unsigned int size, int is_bounded)
 {
 	sync_queue* sq = (sync_queue*) malloc(sizeof(sync_queue));
 	initialize_sync_queue(sq, size, is_bounded);
 	return sq;
 }
 
-void initialize_sync_queue(sync_queue* sq, unsigned long long int size, int is_bounded)
+void initialize_sync_queue(sync_queue* sq, unsigned int size, int is_bounded)
 {
 	pthread_mutex_init(&(sq->q_lock), NULL);
 	pthread_cond_init(&(sq->q_empty_wait), NULL);
